chore: export formatParamUrl
