+++ lib/mode/dynamic.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const { addHook, resolveSwaggerFunction } = require('../util/common')
+const { addHook } = require('../util/add-hook')
+const { resolveSwaggerFunction } = require('../util/resolve-swagger-function')
 
 module.exports = function (fastify, opts, done) {
   opts = Object.assign({}, {
