+++ lib/spec/openapi/utils.js
@@ -1,6 +1,8 @@
 'use strict'
 
-const { readPackageJson, formatParamUrl, resolveLocalRef } = require('../../util/common')
+const { readPackageJson } = require('../../util/read-package-json')
+const { formatParamUrl } = require('../../util/format-param-url')
+const { resolveLocalRef } = require('../../util/resolve-local-ref')
 const { xResponseDescription, xConsume, xExamples } = require('../../constants')
 const { rawRequired } = require('../../symbols')
