+++ lib/spec/swagger/index.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const yaml = require('yaml')
-const { shouldRouteHide } = require('../../util/common')
+const { shouldRouteHide } = require('../../util/should-route-hide')
 const { prepareDefaultOptions, prepareSwaggerObject, prepareSwaggerMethod, normalizeUrl, prepareSwaggerDefinitions } = require('./utils')
 
 module.exports = function (opts, cache, routes, Ref, done) {
