+++ lib/util/resolve-swagger-function.js
@@ -1,0 +1,13 @@
+'use strict'
+
+function resolveSwaggerFunction (opts, cache, routes, Ref, done) {
+  if (typeof opts.openapi === 'undefined' || opts.openapi === null) {
+    return require('../spec/swagger')(opts, cache, routes, Ref, done)
+  } else {
+    return require('../spec/openapi')(opts, cache, routes, Ref, done)
+  }
+}
+
+module.exports = {
+  resolveSwaggerFunction
+}
